Use accent for host config
