+++ packages/component/src/Styles/defaultStyleSetOptions.js
@@ -1,0 +1,19 @@
+const DEFAULT_OPTIONS = {
+  accent: '#69F',
+  avatarSize: 40,
+
+  backgroundColor: '#EEE',
+
+  bubbleBackground: 'White',
+  bubbleImageHeight: 240,
+  bubbleMaxWidth: 480, // screen width = 600px
+  bubbleMinWidth: 250, // min screen width = 300px, Edge requires 372px (https://developer.microsoft.com/en-us/microsoft-edge/platform/issues/13621468/)
+  bubbleTextColor: 'Black',
+
+  scrollToBottomThreshold: 40,
+  sendBoxHeight: 50,
+
+  timestampColor: 'rgba(0, 0, 0, .2)'
+};
+
+export default DEFAULT_OPTIONS
